refactor(ClippingBox)!: move to camera/
the class in question doesn't belong to view/
