+++ lib/src/main/java/io/github/nonmilk/coffee/grinder/camera/PerspectiveCamera.java
@@ -5,7 +5,6 @@ package io.github.nonmilk.coffee.grinder.camera;
 import java.util.Objects;
 
 import io.github.alphameo.linear_algebra.mat.Matrix4;
-import io.github.nonmilk.coffee.grinder.camera.view.ClippingBox;
 import io.github.nonmilk.coffee.grinder.camera.view.PerspectiveView;
 
 /**
